Make the install script idempotent (#11)
* chore: Make script idempotent

* fix: Fix dotfiles path once pulled

* fix: Fix double quotes within double quotes

+++ install.sh
@@ -1,45 +1,66 @@
 #!/bin/sh
 
 # setup `Development` directory
+if ! [ -d "$HOME/Development" ]; then
  mkdir -p ~/Development
+fi
 
 # dotfiles
-cd ~/Development && \
-  git clone https://github.com/civilcode/dotfiles.git
+if ! [ -d "$HOME/Development/dotfiles" ]; then
+  cd ~/Development && \
+    git clone https://github.com/civilcode/dotfiles.git
+else
+  cd ~/Development/dotfiles && \
+   git pull --force
+fi
 
 # rcm
-cd ~/Development && \
-  curl -LO https://thoughtbot.github.io/rcm/dist/rcm-1.3.3.tar.gz && \
-    tar -xvf rcm-1.3.3.tar.gz && \
-    cd rcm-1.3.3 && \
-    ./configure && \
-    make && \
-    sudo make install
+if [ "$(which rcup)" != '/usr/local/bin/rcup' ] ; then
+  cd ~/Development && \
+    curl -LO https://thoughtbot.github.io/rcm/dist/rcm-1.3.3.tar.gz && \
+      tar -xvf rcm-1.3.3.tar.gz && \
+      cd rcm-1.3.3 && \
+      ./configure && \
+      make && \
+      sudo make install
+fi
 
 # docker-compose
-cd ~/Development && \
-  sudo curl -L https://github.com/docker/compose/releases/download/1.21.2/docker-compose-$(uname -s)-$(uname -m) -o /usr/local/bin/docker-compose
-  sudo chmod +x /usr/local/bin/docker-compose
-  docker-compose --version
+if [ "$(which docker-compose)" != '/usr/local/bin/docker-compose' ] ; then
+  cd ~/Development && \
+    sudo curl -L https://github.com/docker/compose/releases/download/1.21.2/docker-compose-$(uname -s)-$(uname -m) -o /usr/local/bin/docker-compose
+    sudo chmod +x /usr/local/bin/docker-compose
+    docker-compose --version
+fi
 
 # hub
-cd ~/Development && \
-  wget https://github.com/github/hub/releases/download/v2.5.0/hub-linux-386-2.5.0.tgz && \
-    tar zvxvf hub-linux-386-2.5.0.tgz && \
-    sudo ./hub-linux-386-2.5.0/install && \
-    rm -rf ./hub-linux-386-2.5.0 && \
-    rm -rf ./hub-linux-386-2.5.0.tgz
+if [ "$(which hub)" != '/usr/local/bin/hub' ] ; then
+ cd ~/Development && \
+   wget https://github.com/github/hub/releases/download/v2.5.0/hub-linux-386-2.5.0.tgz && \
+     tar zvxvf hub-linux-386-2.5.0.tgz && \
+     sudo ./hub-linux-386-2.5.0/install && \
+     rm -rf ./hub-linux-386-2.5.0 && \
+     rm -rf ./hub-linux-386-2.5.0.tgz
+fi
 
 # direnv
-cd ~/Development && \
-  git clone http://github.com/direnv/direnv.git && \
-   cd direnv && \
-   make && \
-   sudo make install
+if [ "$(which direnv)" != '/usr/local/bin/direnv' ] ; then
+  cd ~/Development && \
+    git clone http://github.com/direnv/direnv.git && \
+     cd direnv && \
+     make && \
+     sudo make install
+fi
 
 # setup
 env RCRC=$HOME/Development/dotfiles/rcrc rcup -f && lsrc
-echo "source $HOME/.civilcode.shrc" >> ~/.bashrc
-echo 'eval "$(direnv hook bash)"' >> ~/.shrc.local
+
+if ! grep --quiet "source $HOME/.civilcode.shrc" $HOME/.bashrc;  then
+  echo "source $HOME/.civilcode.shrc" >> ~/.bashrc
+fi
+
+if ! grep --quiet "eval \"$(direnv hook bash)\"" $HOME/.shrc.local;  then
+  echo 'eval "$(direnv hook bash)"' >> ~/.shrc.local
+fi
 
 echo "Finished. Run 'source ~/.bashrc'."
